pygame_pong/requirements.txt
asttokens==3.0.0
colorama==0.4.6
executing==2.1.0
icecream==2.1.3
numpy==2.1.3
pygame-ce==2.5.2
Pygments==2.18.0
PyOpenGL==3.1.7
